LoadBalancing_jxta
The goal of this project is to explore different strategies for load
balancing using JXTA Technology.  JXTA technology is a set of open,
generalized peer-to-peer protocols that allows any connected device on
the network to communicate and collaborate. JXTA is used as the stand
alone technology for peer discovery, implementing multiple peers
advertising a file resource and selecting a server based on algorithms
based on different metrics.